Update public IP code, tests, and godoc comments
Implemented GetPublicIP function so that devs can specify their own
http client (*http.Client). Now using GetPublicIP as a part of the
LogPublicIpAddress convenience function. Also added a tests for
GetPublicIP.

Implemented SimpleJSONResponse as a cleaned up rewrite of
SimpleHttpResponse and deprecated SimpleHttpResponse. Also added a test
for SimpleJSONResponse and reformatted the test for SimpleHttpResponse.

Added and supplemented documentation for all functions and tests,
including some sample code for SimpleJSONResponse